+++ src/machine/turtle.rs
@@ -1,7 +1,7 @@
 pub struct Turtle {
-    pub x: f32,
-    pub y: f32,
-    pub angle: f32,
+    pub x: f64,
+    pub y: f64,
+    pub angle: f64,
 }
 
 impl Turtle {
@@ -13,16 +13,16 @@ impl Turtle {
         }
     }
 
-    pub fn forward(&mut self, d: f32) {
-        self.x += f32::cos(self.angle.to_radians()) * d;
-        self.y += f32::sin(self.angle.to_radians()) * d;
+    pub fn forward(&mut self, d: f64) {
+        self.x += f64::cos(self.angle.to_radians()) * d;
+        self.y += f64::sin(self.angle.to_radians()) * d;
     }
 
-    pub fn left(&mut self, a: f32) {
+    pub fn left(&mut self, a: f64) {
         self.angle -= a;
     }
 
-    pub fn right(&mut self, a: f32) {
+    pub fn right(&mut self, a: f64) {
         self.angle += a;
     }
 }
